Fix compare warnings. (#372)

+++ libpktriggercord/src/pktriggercord-servermode.c
@@ -96,7 +96,7 @@ void write_socket_answer( char *answer ) {
 
 void write_socket_answer_bin( uint8_t *answer, uint32_t length ) {
     ssize_t r = write(client_sock, answer, length);
-    if (r != length) {
+    if ((size_t)r != length) {
         fprintf(stderr, "write(answer) failed");
     }
